[components][moniker] Reduce memory usage
Replace Vec<ChildName> with Arc<[ChildName]> to make monikers cheaper to
clone. Looking at a heapdump of component manager shows that there are
a significant number of clones of Monikers. The ChildNames inside of
Moniker are cheap to clone but the Vec is not. Replacing the Vec with an
Arc removes the allocation from every clone. This has the downside of
making Monikers immutable but Monikers are cloned way more than they are
mutated.

This saves ~2.6MiB on android builds.

Bug: 414632557
Change-Id: I9c0a5189f8d91ae8a06e2645a9e1513908ebf23d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1262564
Reviewed-by: Gary Bressler <geb@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>
Commit-Queue: Chris Drouillard <cdrllrd@google.com>