[super_parser] Update test image to smaller image
Update the test image such that the size is now 10MiB. It was 1 GiB
previously which was unnecessarily large. A large file will be
problematic when we add more test that opens and reads this file as
we'll run out of memory.

Also update the image to be a bit more complex such that we can be more
comprehensive in testing the parser. For example, having sizes for each
partition.

Bug: 404952286
Test: fx test //src/storage/lib/super_parser:super_parser_tests
Change-Id: Ie16dc6d16061897d7f7fcc7fc611b176f156cb9c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1277544
Reviewed-by: Aaron Drew <ripper@google.com>
Commit-Queue: Rui Qi Sim <rqsim@google.com>