[gpt] Re-enable passthrough
This was inadvertently disabled during refactoring in fxrev.dev/1261608.

An integration test for this would be challenging to land, but if
storage-host was enabled on the storage benchmarks, we'd see a hit to
performance benchmarks, which should catch this.

In the meantime, another bug snuck in where we weren't validating
provided offsets.  Change this to happen when they are first provided,
rather than lazily verifying.

Change-Id: I3c4951f06d04df15ee41295c3f19309032352391
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1271864
Reviewed-by: Chris Drouillard <cdrllrd@google.com>
Fuchsia-Auto-Submit: James Sullivan <jfsulliv@google.com>
Commit-Queue: Auto-Submit <auto-submit-builder@fuchsia-internal-service-accts.iam.gserviceaccount.com>