[bt][hfp-hf] Handle unknown AT responses
Log uparseable AT responses and those which the current procedure
can't handle even though they were routed to it.

This should be removed long term and be replaced with dropping the
responsible peer connection, but currently we don't want to do that if
we hit an unimplemented procedure.

Test: N/A
Change-Id: Ib3b65d97185899ec7253e1096e479dc28d5aa6fb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/999610
Reviewed-by: Ani Ramakrishnan <aniramakri@google.com>
Commit-Queue: Matthew Kehrt <kehrt@google.com>