[bt][hfp-hf] Hang up calls.
This commit adds the hang up procedure and call module support for it.

The call module changes are fairly complex.  Determining when we can
stop waiting on a Call's Stream implementation is not simple. We need to
guarantee that we've both received a +CIEV(call = 0) from the peer and
either sent a State::Terminated to the FIDL client or gotten an error
because the client has already closed the channel.

Test: fx test bt-hfp-hands-free-tests
Change-Id: I594d5db7ac4defd165abc419c824c663f26df4e8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1002861
Reviewed-by: Marie Janssen <jamuraa@google.com>
Commit-Queue: Matthew Kehrt <kehrt@google.com>