[fuchsia-async] Change Scope::wake_all to take an active guard
In preparation to switching ExecutionScope to use fuchsia-async's Scope
cancellation API, this changes `wake_all` so that it will take active
guards.  This will allow us to implement `ExecutionScope::shutdown` in
the way that we want.

This also introduces enough of the API to the portable runtime to
support ExecutionScope.

And lastly, we add an on_no_tasks_and_guards which is also needed.

Bug: 416349680
Change-Id: I249ee0f348cdb1a82d22536f999e4a47b98f0c9c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1269644
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Auto-Submit <auto-submit-builder@fuchsia-internal-service-accts.iam.gserviceaccount.com>
Reviewed-by: David Koloski <dkoloski@google.com>
Fuchsia-Auto-Submit: Chris Suter <csuter@google.com>