[tefmocheck] Fix check for Boot-test success string.
For the check that checks whether we find the boot-test successs string
even if the test fails, we need to add exceptStrings to exclude the
string when it's reported in error logs. fxrev.dev/1265145 added the
string to the error message from ReadUntilMatch() so we need to add an
except string for that log as well.

Change-Id: I03041bb115555a9286d6136ebe3283e7f1c70061
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1280484
Reviewed-by: Oliver Newman <olivernewman@google.com>
Commit-Queue: Auto-Submit <auto-submit-builder@fuchsia-internal-service-accts.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Ina Huh <ihuh@google.com>