[vfs] Change ExecutionScope to use fuchsia-async cancellation
There are some changes in semantics:

1. Once shut down and the active-guard count reaches zero, it cannot be
   incremented again, which means that spawning new tasks when a future
   is dropped is no longer guaranteed to always run tasks.  This has
   implications for Fxfs and flushing files. I had to rework the code to
   make sure files get flushed before Fxfs is shut down.

2. Tasks that are spawned no longer hold a reference to its scope,
   which means that the caller must retain the scope.

3. Tasks are no longer guaranteed to run at least once when spawned. We
   no longer seem to depend on this.

Along the way, I spotted an issue with symlinks. They didn't have an
open count so it was possible to unlink them and then you'd get errors
if you tried to read the link. I've changed this to match what happens
on Linux: you can continue to read the link until the symlink is closed.

Bug: 416349680
Change-Id: I6f65cab211cc6b9ccac942bbf4923f1e21b87aa5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1269645
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Chris Suter <csuter@google.com>
Reviewed-by: Claire Gonyeo <cgonyeo@google.com>
Reviewed-by: Sen Jiang <senj@google.com>
Reviewed-by: Brandon Castellano <bcastell@google.com>