[owners] Remove sezers@google.com from src/tests/end_to_end/idle/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.

Please approve this change and vote the highest CR. This will keep the OWNERs
file tidy.  We ask that you do not ignore this change and approve it
unless you know a reason the OWNER should remain.  It can always be
reverted if needed.

If this change is in error, vote the lowest CR value (i.e. reject the CL)
and the bot will abandon it.

See the owner's recent review activity for context:
https://fuchsia-review.googlesource.com/q/sezers@google.com

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: I0e00f1cc0552712dea98359ebe423290923357ae
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1257726
Fuchsia-Auto-Submit: Owners-Cleanup-Prod <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Reviewed-by: Vinicius Felizardo <felizardo@google.com>
Commit-Queue: Auto-Submit <auto-submit-builder@fuchsia-internal-service-accts.iam.gserviceaccount.com>