[assembly] Add repository info to bazel targets by default.
We shouldn't have to explicitly set the repo on every assembly
target. This should automatically set that value in all relevant
assembly bazel targets.

Also provide a way to explicitly set the repository in the bazel target
if necessary (bazelrc flags are not available in every environment that
we want to support).

Bug: b/413433223
Change-Id: I9f51816389779f81d182ae80fd89aab6b7edebcc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1278307
Reviewed-by: Aidan Wolter <awolter@google.com>
Reviewed-by: David Turner <digit@google.com>
Commit-Queue: Jerry Belton <jcecil@google.com>