[socket-proxy] Remove from unified binary
The socket-proxy is not being used on the clients of the unified binary,
so it should be removed. Products that desire to use the socket-proxy
can use the assembly flag to include it in the build.

Update assembly bundles and socket-proxy build targets to only build the
component/package portion of socket-proxy and not the gub portion.

Fixed: 419010176
Change-Id: Idb50d59c3122f4ed25d08a2a7845c150f7338809
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1281087
Commit-Queue: Antonio Linhart <antoniolinhart@google.com>
Reviewed-by: Zoey Bush <zmbush@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>