[owners] Remove danikay@google.com from src/lib/versioning/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.

Please approve this change and vote the highest CR. This will keep the OWNERs
file tidy.  We ask that you do not ignore this change and approve it
unless you know a reason the OWNER should remain.  It can always be
reverted if needed.

If this change is in error, vote the lowest CR value (i.e. reject the CL)
and the bot will abandon it.

See the owner's recent review activity for context:
https://fuchsia-review.googlesource.com/q/danikay@google.com

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: Ic9953626c5aef638cd34f80cd07d21820dd1d043
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1276772
Reviewed-by: Hunter Freyer <hjfreyer@google.com>
Commit-Queue: Auto-Submit <auto-submit-builder@fuchsia-internal-service-accts.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Owners-Cleanup-Prod <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>