[starnix] Assert that boot-to-UTC ref ticks is not zero
Assert that boot-to-UTC reference ticks is never zero when updating the
UTC transform in the starnix kernel vdso. Setting this value to zero
will cause any process that calculates UTC time using the vdso to
assert.

This is not necessarily indicative of the root cause of the issue
described in the bug, but it is a reasonable defense-in-depth measure to
catch mistakes in the calculation or propagation of this sensitive
value.

Bug: b/415099415
Change-Id: If6a79cf5ff13464dad2e753548b01b98bca3973d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1265841
Commit-Queue: Corey Tabaka <eieio@google.com>
Reviewed-by: James Robinson <jamesr@google.com>