[libc] Rework ThreadAllocator into ThreadStorage
ThreadAllocator was originally designed as a temporary object
that owns the storage only during the setup phase and does not
record all the information necessary to use that storage.  It now
seems more useful to have a long-lived ThreadStorage object that
is optimized for compactly storing all the necessary information.
For now, the ownership is moved between ThreadStorage and the
legacy C `struct pthread`.  When that's replaced by a proper C++
Thread object, it will just have a ThreadStorage member instead
to track the ownership over the whole lifetime of the thread.

Bug: 342469121
Change-Id: I12e864ac059f4f2c8be9e342518a12dce90fa873
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1275705
Fuchsia-Auto-Submit: Roland McGrath <mcgrathr@google.com>
Reviewed-by: Paul Kirth <paulkirth@google.com>
Commit-Queue: Auto-Submit <auto-submit-builder@fuchsia-internal-service-accts.iam.gserviceaccount.com>