[rust][3p] Switch to use rustc_library
... instead of `rust_library`.

Motivation of this change is to reuse the logic for defining rustdoc
subtargets from `rustc_library`. The same logic was previously repeated
in BUILDCONFIG.gn for rust_library, making them hard to stay consistent.

NOTE: cargo-gnaw unfortunately has to pick up details about
`rustc_library` (field names and the .actual subtarget) in this change,
which means it will be coupled with the Fuchsia build after this change.

Also note: The changes in `icu_rust_crates.gni` are enough to make the
build pass, this file will be removed in a followup to simplify diff,
where `icu_rustc_library` is used instead of `icu_rust_library`.

Bug: 355454632

Change-Id: Ie2a9c3138277c8e032f067aa9392f6660d074e5a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1277545
Commit-Queue: Jay Zhuang <jayzhuang@google.com>
Reviewed-by: Erick Tryzelaar <etryzelaar@google.com>