[blobfs] Fixes handle close race for debug assert.
Outside of debug mode the response would still have been correct, but it
was via an unexpected code path. If a handle closure for the writer
hasn't been processed before the next open attempt then we followed a
code path that assumed we were still writing the blob but it had been
closed by the client.

The new case confirms that the writer is complete and the blob is
persistent, which obsoletes the previous "else" case below.

Bug: 417273110
Change-Id: Ib23056ad5fcfe4742416862e7df698824c135308
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1281387
Commit-Queue: Martin Lindsay <mlindsay@google.com>
Reviewed-by: Chris Drouillard <cdrllrd@google.com>