[fuchsia-async] Forbid wakers from changing lists
In theory, it was possible for a WakerEntry to change to a different
list in which case, bad things would happen.  This change prevents that
at the cost of having to always acquire the lock when the WakerEntry is
dropped.

Change-Id: I0c6d46cd812236da819036b828d72da8fa50df60
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1281845
Reviewed-by: Christopher Johnson <crjohns@google.com>
Commit-Queue: Auto-Submit <auto-submit-builder@fuchsia-internal-service-accts.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Chris Suter <csuter@google.com>