[starnix][trace] Don't alloc DurationScopes when tracing is off
Currently, starnix's trace_duration macro checks if tracing is compiled
in, and if so, always allocates a Duration Scope, regardless of if
tracing is actually enabled or not. We should avoid this and only create
one if we are actually tracing.

Change-Id: I2c9e9e86393f591d5fdae873721141986ab77085
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1280149
Fuchsia-Auto-Submit: Gwen Mittertreiner <gmtr@google.com>
Commit-Queue: Auto-Submit <auto-submit-builder@fuchsia-internal-service-accts.iam.gserviceaccount.com>
Reviewed-by: David Song <wintermelons@google.com>
Reviewed-by: Fadi Meawad <fmeawad@google.com>