[wlan][policy] Adds PHY reset control
This change enables wlancfg to reset WLAN PHYs.  This code
path is triggered in build configurations where thresholded
recovery is enabled and when WLAN is detected to be in a
degraded state.  Specifically, the reset can trigger if a
device experiences repeated failures to connect to a
network, scan for networks, start an access point, create
interfaces, or destroy interfaces.

Test: fx test wlancfg-tests
Multiply: wlancfg-tests
Bug: b/328463165
Change-Id: I6fcd53fad23e98b3af24554a721117dece448da1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/987896
Reviewed-by: Naomi Mccracken <nmccracken@google.com>
Commit-Queue: Kevin Sakuma <sakuma@google.com>