[test-pilot] update output directory/outcome
This CL changes the directory structure of the output directory used
by test-pilot to conform with the new design. It also updates the
outcome type used in the summary to have a 'result' field and 'detail'
field, also in conformance with the current design.

Test: new and updated unit tests

Bug: 327640651
Change-Id: I2ad957f1772ce6fedba25f65e35f69a20033b64e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1281590
Commit-Queue: Dale Sather <dalesat@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>