[block] Return errors on bad alignment instead of asserting.
Change the block client to stop asserting on bad alignment arguments and
moves to returning INVALID_ARGS instead. Adds tests for the cases as
well.

Leaves FakeDevice as it is a testing application, and failing loudly is
better there, and leaving the file-backed device implementation as it is
specified to only build for host.

Bug: 415893572
Change-Id: Iea5341a92788ded70e2dcf7a4e7076d3c9917b7c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1282625
Reviewed-by: Aaron Drew <ripper@google.com>
Commit-Queue: Martin Lindsay <mlindsay@google.com>