[fuchsia-async] Support half-closed zircon sockets
Add support for waking up wakers when a zircon socket is half-closed on
fuchsia_async::Socket. The existing implementation would haphazardly
return BAD_STATE if it manages to observe the half-closed state, or pend
forever if it doesn't. Given no systems currently reopen half-closed
sockets the default behavior of loudly bailing out of waiting sounds
like the right one.

Add an escape hatch for a socket that ignores half-closed, and pends
until either readable or writable are available. The plan is to use this
version in overnet (pending addition of emulated support for this) so it
can proxy the half-closed signals.

Change-Id: I1d41921cee5744f46d8cbd1eec7f40f0b8381aab
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1280486
Reviewed-by: Gary Bressler <geb@google.com>
Fuchsia-Auto-Submit: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Auto-Submit <auto-submit-builder@fuchsia-internal-service-accts.iam.gserviceaccount.com>
Reviewed-by: Chris Suter <csuter@google.com>