[pkg] Merge full-resolver with pkg-resolver
In Fuchsia, component resolution is composed of two layers: the
component resolver, and the package resolver. Both Component Framework
and Software Delivery define their own resolver abstraction.

Currently, there exists an independent component (full-resolver.cm) that
adapts the package resolver to the component resolver ABI. However, the
implementation of this component is quite simple and there isn't much of
a compelling reason to keep it separate. For that reason, this commit merges
full-resolver with pkg-resolver, allowing us to save a bit of memory.

The full-resolver implementation is not changed in the commit. As a
followup, we can simply the implementation to call the pkg-resolver rust
routines directly, instead of through fuchsia.pkg.PackageResolver
protocol (which now just loops back to the same component).

 ## Resource footprint

This change saves memory on eng builds by removing full-resolver from
core. Also, the instance of full-resolver in bootstrap for driver
loading is replaced with loopback-resolver, which is considerably smaller
(it's a simple C++ program). In non-eng builds, at worst case this adds
a negligible amount of memory footprint to pkg-resolver, which now has
the code and data that previously belonged to full-resolver. If the full
resolver service is never used, the data cost is zero.

Bug: 417525026

Change-Id: Ief27269d20bddf609add45b03520cb4e68ca191f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1274091
Reviewed-by: Ben Keller <galbanum@google.com>
Reviewed-by: Laura Peskin <pesk@google.com>
Commit-Queue: Gary Bressler <geb@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: Aidan Wolter <awolter@google.com>