[userboot] Wait for bootstrap channel to become readable
The userboot code never waited for its channel to become
readable.  In the past, the kernel stuffed the message into the
channel before starting the process, so this was overlooked for
years.  Now the process starts running in parallel with the
kernel stuffing the bootstrap message, so it became a race.  The
race is probably an optimization so the userboot startup work
before the channel read starts earlier.  But userboot needs to
wait in case it got there first.

Bug: 416531428
Change-Id: I9e086b84c60e34eeb0f4050dde174ab8c7a017b4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1283404
Reviewed-by: Gianfranco Valentino <gevalentino@google.com>
Fuchsia-Auto-Submit: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Auto-Submit <auto-submit-builder@fuchsia-internal-service-accts.iam.gserviceaccount.com>