[testing] test attribute on host no longer initializing logging
The docs for the macro #[fuchsia::test] state that when implemented on
the host, they do not initialize a logger and that the host must have
logging initialized. That was not true until this CL.

This adds tests to ensure that when a #[fuchsia::test] is run on the
host that a logger is not initialized, and that when it is run on the
target that it is initialized.

Change-Id: Iebbea6b4573f72ba2b02de8cbcf14ede6a6a636c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1283246
Commit-Queue: Colin Nelson <colnnelson@google.com>
Reviewed-by: Marie Janssen <jamuraa@google.com>
Fuchsia-Auto-Submit: Colin Nelson <colnnelson@google.com>