[component_manager] Optimize path representation
Before this change, all `cm_types::Path*` types stored the path as a
list of `Name` segments. This involves two levels of dynamic allocation:
one for the `Vec` and another for the `Name`, which is backed by
`FlyStr`.

- An additional 64 bits (one pointer) to store each path segment
- 64 bits to store the size of each segment (although this is mitigated
  by flyweights)
- Internal fragmentation from Vec capacity that goes unused

It is more space efficient to store the path as a `/`-separated string
-- whose overhead is only (n - 1) bytes where n is the number of path
segments.

Note that this change requires adding a small amount of complexity to
the type system. There is now a new type `NameRef`, an unowned
variant of `Name`, that is used to represent a path segment. Before we
used `&Name` to do this job, which was possible because each path
segment was stored as a `Name`. To avoid allocating a `Name`, we
introduce the unowned type which is backed by a `&str` slice.

Appears to save ~600KiB on release builds.

Change-Id: Ib3d713fa147a6599faccbc856b5243bc88f8ecc8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1278366
Commit-Queue: Gary Bressler <geb@google.com>
Reviewed-by: Justin Mattson <jmatt@google.com>
Reviewed-by: Claire Gonyeo <cgonyeo@google.com>