[rust][zx] Make `zx::Counter::create()` infallible
This change modifies the return type signature to align with the current
APIs of other Rust kernel wrapper objects in the fuchsia runtime, such
as `zx::Event::create()` and `zx::EventPair::create()`, all of which
return `Self` and not `Result`.

This is a marginal improvement in the quality of life for the users,
until such a time that we re-acquire the need for fallible constructor
functions.

Depends-on: turquoise-internal:I15126b00c2624ff42a53148d04121b50b1555569
Change-Id: Idda452e485800651dd2e8e4b7e14adbd67cb4bb2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1283890
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Reviewed-by: Adam Barth <abarth@google.com>
Fuchsia-Auto-Submit: Filip Filmar <fmil@google.com>
Dependencies-Satisfied: CL Deps Checker <cl-deps-checker@fuchsia-internal-service-accts.iam.gserviceaccount.com>
Commit-Queue: Auto-Submit <auto-submit-builder@fuchsia-internal-service-accts.iam.gserviceaccount.com>