[wlan][policy] fix roam on wpa1wpa2 bug
There was a bug where if a network was saved as WPA2 but the actual
security type of the AP being used was Wpa1Wpa2Personal, roaming would
not consider roaming to APs with the same security type. This was
because WPA1WPA2Personal was not in the match case for networks
compatible to WPA2. This also adds e2e test cases for some of the
security permutations including this one.

Bug: 417771480
Test: added unit tests: fx test wlancfg-tests
Change-Id: I7944866acc16c8e4450b992d05affc5aa95c5a6e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1281584
Reviewed-by: Hayden Nix <haydennix@google.com>
Commit-Queue: Naomi Mccracken <nmccracken@google.com>