[devicetree][arm-gic] Implement interrupt wake vector support
Add support for fuchsia,interrupt-wake-vector for specifying interrupts
that should have the ability to wake the system from suspend.

Eventually this should become part of InterruptParser, but implementing
it in the gic visitor as a first pass allows us to migrate and test it
before worrying about how to make the interface generic.

Bug: https://fxbug.dev/402466083
Change-Id: I5f866d1c9cca028e460c7c006bdcc2fa6902fc6d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1277163
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: Puneetha Ramachandra <puneetha@google.com>
Commit-Queue: Christopher Anderson <cja@google.com>