[fxfs] Fix FxVolume terminate hang when read-only
When we mount a filesystem with the read-only option, we never start the
reap task. This was problematic for FxVolume::terminate(), which
attempts to flush the graveyard (which will never complete due to no
reap task handling it). When the filesystem is read-only, we now avoid
doing any work that attempts to mutate the filesystem during volume
termination.

Likewise, when unlocking an encrypted store, we also make sure that
they must be unlocked as read-only if the filesystem itself is
read-only.

Test: fx test fxfs-platform-test
Change-Id: I165b7558fdf81040f628e56e9fae74932366c4c3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1284684
Reviewed-by: James Sullivan <jfsulliv@google.com>
Commit-Queue: Brandon Castellano <bcastell@google.com>