[kernel][vm] Remove mapped and unmapped out args
These out arguments have ended up being redundant in the present
implementation, with all architectures either doing the full operation,
or doing nothing. This is reflected in how the set the the value of
this out parameter, with it just being set to the passed in |count|,
making any error checking or other logic against the value pointless.

Change-Id: I2ed9bbe2e0ff782b0d56dff8081a9a43bd78e2e8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1280825
Commit-Queue: Adrian Danis <adanis@google.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>