[assembly] Boot tests modify the board to insert new partitions
Instead of passing the partitions config to the product_bundle, we
are moving this to be specified by the board. Because boot tests use
a custom partitions config with the EFI, we have to modify the board
to use the correct partitions config.

Change-Id: Id07bea8504c91b4b77a72396b91c6924ff8fec2e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1285984
Fuchsia-Auto-Submit: Aidan Wolter <awolter@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>
Commit-Queue: Auto-Submit <auto-submit-builder@fuchsia-internal-service-accts.iam.gserviceaccount.com>