[netstack3] Internal SocketCookie support for SO_COOKIE and eBPF
Previously netstack3_sync contained debugId which was also used to
produce TraceResourceId. The same value can be reused as a "socket
cookie" to identify sockets outside of netstack3.

Added `ResourceToken` that contains a unique 64-bit value for all
ref-counted resources. It's used now to produce `DebugId` and
`TraceResourceId` for ref-counted objects. It's also used in to produce
`SocketCookie` and pass it to eBPF programs.

Change-Id: I066b9359c74a4c1b8a2e5d2468c03b06d2134d57
Bug: 406296773
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1280161
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Sergey Ulanov <sergeyu@google.com>