[mm2] Do not postpone captures on pressure changes.
This fixes an issue where captures could be arbitrarily deferred in
some cases.

memory_monitor2 does regular memory usage captures based on the
current memory pressure. When the memory pressure changes, the
frequency changes as well (so that it's possible to look at memory
usage more frequently when the system is under load, to identify what
takes so much memory, for instance).

Prior to this change, a memory pressure change would reset the timer,
so that if memory pressure changes occurred more frequently than the
configured frequency, we would never do any capture. Even when
captures are not postponed, they are captured haphazardly.

With this change, we never delay the next capture, even if the
pressure changes.

Bug: 417722087
Change-Id: Ia60751d68f333a629af60f998636baf212c7aef2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1285624
Reviewed-by: Étienne J. Membrives <etiennej@google.com>
Commit-Queue: Guillaume Chelfi <chelfi@google.com>