[build][idk] Remove write_stable_true_to_meta_json
https://fxrev.dev/1246507 incorrectly added
add_stable_true_to_meta_json` to `sdk_atom()` as a replacement for
`add_stable_property`, but the latter was use for the _build_ manifest,
not the `meta.json` file, which always contains "stable" for atom types
that support it.

https://fxrev.dev/1078492 then added `write_stable_true_to_meta_json`
based on the value of `add_stable_property` but always used it as the
value for "stable" in the meta.json file. This can be achieved using
the value of "is_stable" in the prebuild info instead.

Bug: 338009514
Test: `fx build //sdk:in_tree_collection && build/sdk/generate_prebuild_idk/generate_prebuild_idk.sh --prebuild-manifest $(fx get-build-dir)/sdk/prebuild/in_tree_collection.json --build-dir $(fx get-build-dir) --check` passes.
Change-Id: I91f59019a2350c8beb514493e45019deaf6922b3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1286430
Fuchsia-Auto-Submit: David Dorwin <ddorwin@google.com>
Fuchsia-Auto-Submit: David Turner <digit@google.com>
Reviewed-by: David Turner <digit@google.com>
Commit-Queue: David Dorwin <ddorwin@google.com>