[dhcp-client] Allow Enobufs to be observed
Allow Enobufs to be observed on packet sockets only, given this is
observable if the tx queue of the device is full. Even if backpressure
_was_ implemented for packet sockets, the back-pressure is per socket,
while the nobufs error is observed if a device itself is overwhelmed.

Fixed: 420719068
Change-Id: I00566c359f39a4c0a5de979b44ce12b90124c0fb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1286767
Fuchsia-Auto-Submit: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Evan Wildenhain <wildenhain@google.com>