[fshost] add publisher matcher
Instead of publishing unmanaged, unmatched block devices to the block
directory in the manager, add a matcher that is put at the end of the
other matchers as a catch-all. This captures devices that aren't matched
by anything else implicitly, and keeps the manager loop cleaner.

Other matchers that can publish will be added so it made sense to
refactor this.

Bug: 407072932
Change-Id: I37552981128ac06982edc894cdfe9c7f692f88b1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1257991
Fuchsia-Auto-Submit: Stephen Demos <sdemos@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>
Commit-Queue: Stephen Demos <sdemos@google.com>