[starnix][trace] Correcting newlines to for ATrace events
This checks fror a  newline character  at the end of the ATRACE data.
Since the TraceEvent struct appends a newline unconditionally,the
newline is removed before passing the data to TraceEvent in order to
avoid creating empty records when reading the event data.

This also sets the common kernel pid to be the thread id from Linux,
which reduces the collision of the events on the main thread.

Bug: 401645216
Change-Id: Ie2836e7cccb9ecab459acf2a7bd224675df1a196
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1286846
Reviewed-by: David Song <wintermelons@google.com>
Reviewed-by: Jia Li <lizhijia@google.com>
Reviewed-by: Gwen Mittertreiner <gmtr@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>