[starnix] Remove lifetime parameter on Locked
The lifetime parameter on the &mut that Locked is always used with is
essentially redundant with the type parameter. Same safety, slightly
less weird sigils!

There is one line of unsafe code, which creates a &mut Locked<T> out of
nothing whatsoever. This is safe because Locked is a ZST.

Another option is to remove the `&mut` part but keep the lifetime
parameter. (We need to have at least one to maintain safety, because
that's what stops you from saving the token somewhere and using it
later.) There are two marginal downsides to this:

1. <'_ is somewhat more awkward to type than &mut.
2. It would be necessary to write a function call every time you pass
   the Locked token down.

Change-Id: If272ff71ea76120d5a5749eb71b28a082a43f9e1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1286765
Commit-Queue: Theodore Dubois <tbodt@google.com>
Reviewed-by: Adam Barth <abarth@google.com>
Fuchsia-Auto-Submit: Theodore Dubois <tbodt@google.com>