[owners] Remove neelsa@google.com from sdk/lib/input_report_reader/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.

Please approve this change and vote the highest CR. This will keep the OWNERs
file tidy.  We ask that you do not ignore this change and approve it
unless you know a reason the OWNER should remain.  It can always be
reverted if needed.

If this change is in error, vote the lowest CR value (i.e. reject the CL)
and the bot will abandon it.

See the owner's recent review activity for context:
https://fuchsia-review.googlesource.com/q/neelsa@google.com

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: Ic404b667c9b0be1d57ac993cce2d88be70d212f1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1276751
Commit-Queue: Ruby Zhuang <rdzhuang@google.com>
Fuchsia-Auto-Submit: Ruby Zhuang <rdzhuang@google.com>
Reviewed-by: Ruby Zhuang <rdzhuang@google.com>