[security] Allowlist addition for new guest manager
The InteractiveDebianGuest is a wrapper around the GuestManager, but
notably provides some additional functionality and interactions. The
intent will be to use this virtualized environment for Starnix testing,
so that we can run tests against both Starnix and arbitrary Linux
kernels. As such, added the moniker to the relevant policies.

Bug: 418575059
Test: build and run with a relevant component topology
Change-Id: Ie373b107ed5ae6c5296a17b6e8a7bc696114df4f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1287337
Fuchsia-Auto-Submit: Aaron Knobloch <aknobloch@google.com>
Commit-Queue: Aaron Knobloch <aknobloch@google.com>
Reviewed-by: Cameron Finucane <eep@google.com>