[starnix][vfs] Make FsNode::node_id immutable
Previously, we set the node_id when we installed a directory as the root
of a file system. This CL refactors the code is that we figure out the
node_id for the root of the file system before we create the FsNode.

This CL also makes FsNode::fs immutable.

Bug: b/419306849
Change-Id: Ie3d2842e11b2dc97303e11b42b179949e59b66d5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1288248
Commit-Queue: Adam Barth <abarth@google.com>
Reviewed-by: Jeremiah Griffin <jzgriffin@google.com>
Reviewed-by: James Robinson <jamesr@google.com>
Fuchsia-Auto-Submit: Adam Barth <abarth@google.com>