[dl] Support NULL & empty string dlopen arguments
This CL adds support for libdl to return a handle to the executable
module when NULL or an empty string is passed to dlopen.

A simple test is added to verify the module pointer returned by dlopen
is for the executable, by introducing the `ModuleLinkMap` function to
the DlImpl and System test fixtures.

Changes were made to the test fixture tracking function `TrackModule`
to use "<null>" as the given name for a handle representing the NULL
pointer file argument.

Fixed: 369684039, 361674544
Change-Id: I086581591c2fc1d01c2aedac8c8047590c1f9e0f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1288547
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Caslyn Tonelli <caslyn@google.com>
Fuchsia-Auto-Submit: Caslyn Tonelli <caslyn@google.com>